+++ src/app/core/@ngrx/products/products.facade.ts
@@ -5,14 +5,15 @@ import { Store, select } from '@ngrx/store';
 import {
   selectProductsData,
   selectProductsError,
-  selectProductsLoaded
+  selectProductsLoaded,
+  selectSelectedProductByUrl
 } from './products.selectors';
 import * as ProductsActions from './../../../core/@ngrx/products/products.actions';
 
 // rxjs
 import { Observable } from 'rxjs';
 
-import { IProduct } from './../../../products/models/product.model';
+import { IProduct, Product } from './../../../products/models/product.model';
 
 @Injectable({
   providedIn: 'root'
@@ -21,11 +22,13 @@ export class ProductsFacade {
   products$: Observable<ReadonlyArray<IProduct>>;
   productsError$: Observable<Error | string>;
   productsLoaded$: Observable<boolean>;
+  selectedProductByUrl$: Observable<Product>;
 
   constructor(private store: Store) {
     this.productsError$ = this.store.select(selectProductsError);
     this.productsLoaded$ = this.store.select(selectProductsLoaded);
     this.products$ = this.store.select(selectProductsData);
+    this.selectedProductByUrl$ = this.store.select(selectSelectedProductByUrl);
   }
 
   loadProducts(): void {