+++ src/app/core/@ngrx/root-store.module.ts
@@ -8,13 +8,26 @@ import { StoreDevtoolsModule } from '@ngrx/store-devtools';
 import { environment } from './../../../environments/environment';
 
 import { EffectsModule } from '@ngrx/effects';
+import { CustomSerializer, RouterEffects, routerReducers } from './router';
+import { RouterState, StoreRouterConnectingModule } from '@ngrx/router-store';
 
 @NgModule({
   declarations: [],
   imports: [
     CommonModule,
-    StoreModule.forRoot({}),
-    EffectsModule.forRoot([]),
+    StoreModule.forRoot(routerReducers, {
+      runtimeChecks: {
+        strictStateSerializability: true,
+        strictActionWithinNgZone: true,
+        strictActionTypeUniqueness: true
+      }
+    }),
+    StoreRouterConnectingModule.forRoot({
+      stateKey: 'router',
+      routerState: RouterState.Minimal,
+      serializer: CustomSerializer
+    }),
+    EffectsModule.forRoot([RouterEffects]),
     !environment.production ? StoreDevtoolsModule.instrument() : [],
     ProductsStoreModule,
   ]