+++ changelog.md
@@ -84,3 +84,4 @@ Added ngrx; Added Products store
 
 06.04.2021
 Connect Products store with components via facade
+Added Router store